requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.2
pdfminer.six==20201018
pipenv==2020.11.15
protobuf==3.13.0
pycparser==2.20
pycryptodome==3.9.9
pylint==2.5.3
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.1
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xlrd==1.2.0
